refactor: unify page title API into <PageTitle> and colocate layout primitives
Collapse the dual <Title>/useDeclareAppHeaderTitle call pattern into a single
<PageTitle> that sets both the <title> element and the app-header store. Move
page-chrome primitives (PageTitle, PageContainer, title store) under
src/app/layouts/. Fixes a latent bug on board-page where the loading branch
returned before <Title> rendered, leaving the browser tab stale.
