+++ src/app/layouts/page-title.tsx
@@ -1,10 +1,12 @@
 import { APP_NAME } from "@app/app-info";
+import { useSetPageTitle } from "./page-title-store";
 
-export interface TitleProps {
+export interface PageTitleProps {
   children: string | undefined;
 }
 
-export function Title({ children }: TitleProps) {
+export function PageTitle({ children }: PageTitleProps) {
+  useSetPageTitle(children);
   if (!children) {
     return null;
   }
